+++ examples/simple.rs
@@ -597,10 +597,8 @@ impl SimpleWindow {
         button1.set_color(CONFIG_TEXT_COLOR, [0.0, 0.0, 0.0, 1.0]);
         button1.set_numeric(CONFIG_BORDER_WIDTH, 2);
         button1.set_color(CONFIG_BORDER_COLOR, [0.0, 0.0, 0.0, 1.0]);
-        button1.on_click(|x, widgets| {
-            let state = widgets[get_widget_position_by_name(widgets, "BoxInLayoutWidget1".to_string()) as usize]
-                .widget
-                .borrow_mut()
+        button1.get_callbacks().on_click(|x, widgets| {
+            let state = get_widget_by_name(widgets, "BoxInLayoutWidget1".to_string())
                 .config()
                 .get_toggle(CONFIG_WIDGET_HIDDEN);
             let button_text = if state == true {
@@ -611,9 +609,7 @@ impl SimpleWindow {
 
             x.set_config(CONFIG_DISPLAY_TEXT, Config::Text(button_text));
 
-            widgets[get_widget_position_by_name(widgets, "BoxInLayoutWidget1".to_string()) as usize]
-                .widget
-                .borrow_mut()
+            get_widget_by_name(widgets, "BoxInLayoutWidget1".to_string())
                 .set_toggle(CONFIG_WIDGET_HIDDEN, !state);
 
             invalidate_all_widgets_except(widgets, x.get_widget_id());
@@ -637,10 +633,8 @@ impl SimpleWindow {
         button3.set_color(CONFIG_TEXT_COLOR, [0.0, 0.0, 0.0, 1.0]);
         button3.set_numeric(CONFIG_BORDER_WIDTH, 2);
         button3.set_color(CONFIG_BORDER_COLOR, [0.0, 0.0, 0.0, 1.0]);
-        button3.on_click(|x, widgets| {
-            let state = widgets[get_widget_position_by_name(widgets, "BoxInLayoutWidget2".to_string()) as usize]
-                .widget
-                .borrow_mut()
+        button3.get_callbacks().on_click(|x, widgets| {
+            let state = get_widget_by_name(widgets, "BoxInLayoutWidget2".to_string())
                 .config()
                 .get_toggle(CONFIG_WIDGET_DISABLED);
             let button_text = if state == true {
@@ -651,29 +645,19 @@ impl SimpleWindow {
 
             x.set_config(CONFIG_DISPLAY_TEXT, Config::Text(button_text));
 
-            widgets[get_widget_position_by_name(widgets, "BoxInLayoutWidget2".to_string()) as usize]
-                .widget
-                .borrow_mut()
+            get_widget_by_name(widgets, "BoxInLayoutWidget2".to_string())
                 .set_toggle(CONFIG_WIDGET_DISABLED, !state);
 
-            widgets[get_widget_position_by_name(widgets, "MiniBox1".to_string()) as usize]
-                .widget
-                .borrow_mut()
+            get_widget_by_name(widgets, "MiniBox1".to_string())
                 .set_toggle(CONFIG_WIDGET_DISABLED, !state);
 
-            widgets[get_widget_position_by_name(widgets, "MiniBox2".to_string()) as usize]
-                .widget
-                .borrow_mut()
+            get_widget_by_name(widgets, "MiniBox2".to_string())
                 .set_toggle(CONFIG_WIDGET_DISABLED, !state);
 
-            widgets[get_widget_position_by_name(widgets, "MiniBox3".to_string()) as usize]
-                .widget
-                .borrow_mut()
+            get_widget_by_name(widgets, "MiniBox3".to_string())
                 .set_toggle(CONFIG_WIDGET_DISABLED, !state);
 
-            widgets[get_widget_position_by_name(widgets, "MiniBox4".to_string()) as usize]
-                .widget
-                .borrow_mut()
+            get_widget_by_name(widgets, "MiniBox4".to_string())
                 .set_toggle(CONFIG_WIDGET_DISABLED, !state);
         });
 
@@ -695,21 +679,16 @@ impl SimpleWindow {
         button5.set_color(CONFIG_TEXT_COLOR, [0.0, 0.0, 0.0, 1.0]);
         button5.set_numeric(CONFIG_BORDER_WIDTH, 2);
         button5.set_color(CONFIG_BORDER_COLOR, [0.0, 0.0, 0.0, 1.0]);
-        button5.on_click(|x, widgets| {
-            let bilw3 = get_widget_position_by_name(widgets, "BoxInLayoutWidget3".to_string());
-
-            widgets[bilw3 as usize]
-                .widget
-                .borrow_mut()
-                .set_config(
-                    CONFIG_MAIN_COLOR,
-                    Config::Color([
-                        (rand::random::<u8>() as f32 / 255.0),
-                        (rand::random::<u8>() as f32 / 255.0),
-                        (rand::random::<u8>() as f32 / 255.0),
-                        1.0,
-                    ]),
-                );
+        button5.get_callbacks().on_click(|x, widgets| {
+            get_widget_by_name(widgets, "BoxInLayoutWidget3".to_string()).set_config(
+                CONFIG_MAIN_COLOR,
+                Config::Color([
+                    (rand::random::<u8>() as f32 / 255.0),
+                    (rand::random::<u8>() as f32 / 255.0),
+                    (rand::random::<u8>() as f32 / 255.0),
+                    1.0,
+                ]),
+            );
         });
 
         self.pushrod.borrow_mut().add_widget_to_layout_manager(
@@ -811,21 +790,16 @@ impl SimpleWindow {
         button2.set_color(CONFIG_TEXT_COLOR, [0.0, 0.0, 0.0, 1.0]);
         button2.set_numeric(CONFIG_BORDER_WIDTH, 2);
         button2.set_color(CONFIG_BORDER_COLOR, [0.0, 0.0, 0.0, 1.0]);
-        button2.on_click(|x, widgets| {
-            let widget_id = get_widget_position_by_name(widgets, "ProgressWidget".to_string());
-
-            widgets[widget_id as usize]
-                .widget
-                .borrow_mut()
-                .set_config(
-                    CONFIG_SECONDARY_COLOR,
-                    Config::Color([
-                        (rand::random::<u8>() as f32 / 255.0),
-                        (rand::random::<u8>() as f32 / 255.0),
-                        (rand::random::<u8>() as f32 / 255.0),
-                        1.0,
-                    ]),
-                );
+        button2.get_callbacks().on_click(|x, widgets| {
+            get_widget_by_name(widgets, "ProgressWidget".to_string()).set_config(
+                CONFIG_SECONDARY_COLOR,
+                Config::Color([
+                    (rand::random::<u8>() as f32 / 255.0),
+                    (rand::random::<u8>() as f32 / 255.0),
+                    (rand::random::<u8>() as f32 / 255.0),
+                    1.0,
+                ]),
+            );
         });
 
         self.pushrod.borrow_mut().add_widget_to_layout_manager(