+++ src/widget/image_widget.rs
@@ -18,6 +18,8 @@ use opengl_graphics::{GlGraphics, Texture, TextureSettings};
 
 use crate::widget::config::*;
 use crate::widget::widget::*;
+use crate::core::callbacks::*;
+use crate::core::widget_store::*;
 
 /// Draws an image.
 pub struct ImageWidget {
@@ -25,6 +27,7 @@ pub struct ImageWidget {
     image: Texture,
     image_size: crate::core::point::Size,
     widget_id: i32,
+    callbacks: DefaultWidgetCallbacks,
 }
 
 impl ImageWidget {
@@ -47,6 +50,7 @@ impl ImageWidget {
                 h: texture_height,
             },
             widget_id: 0,
+            callbacks: DefaultWidgetCallbacks::new(),
         }
     }
 }
@@ -83,6 +87,22 @@ impl Widget for ImageWidget {
         self.widget_id
     }
 
+    fn handle_event(
+        &mut self,
+        injected: bool,
+        _event: CallbackEvent,
+        _widget_store: Option<&Vec<WidgetContainer>>,
+    ) -> Option<CallbackEvent> {
+        if !injected {
+        }
+
+        None
+    }
+
+    fn handles_events(&mut self) -> bool {
+        true
+    }
+
     fn get_injectable_custom_events(&mut self) -> &mut dyn InjectableCustomEvents {
         self
     }
@@ -93,5 +113,9 @@ impl Widget for ImageWidget {
 
     fn get_drawable(&mut self) -> &mut dyn Drawable {
         self
+    }
+
+    fn get_callbacks(&mut self) -> &mut DefaultWidgetCallbacks {
+        &mut self.callbacks
     }
 }