+++ src/widget/push_button_widget.rs
@@ -33,7 +33,7 @@ pub struct PushButtonWidget {
     text_widget: TextWidget,
     active: bool,
     widget_id: i32,
-    on_click: Option<Box<dyn FnMut(&mut PushButtonWidget, &Vec<WidgetContainer>)>>,
+    callbacks: DefaultWidgetCallbacks,
 }
 
 impl PushButtonWidget {
@@ -51,7 +51,7 @@ impl PushButtonWidget {
             text_widget,
             active: false,
             widget_id: 0,
-            on_click: None,
+            callbacks: DefaultWidgetCallbacks::new(),
         }
     }
 
@@ -67,26 +67,6 @@ impl PushButtonWidget {
         self.text_widget
             .set_color(CONFIG_TEXT_COLOR, [0.0, 0.0, 0.0, 1.0]);
         self.invalidate();
-    }
-
-    /// Sets a callback closure that can be called when a click is registered for this
-    /// widget.
-    pub fn on_click<F>(&mut self, callback: F)
-    where
-        F: FnMut(&mut PushButtonWidget, &Vec<WidgetContainer>) + 'static,
-    {
-        eprintln!("Setting on click callback.");
-        self.on_click = Some(Box::new(callback));
-    }
-
-    /// Calls the click `on_click` callback, if set.  Otherwise, ignored.  Sends a reference
-    /// of the current `Widget` object as a parameter, so this object can be modified when
-    /// a click is registered, if necessary.
-    pub fn click(&mut self, widgets: &Vec<WidgetContainer>) {
-        if let Some(mut cb) = self.on_click.take() {
-            cb(self, widgets);
-            self.on_click = Some(cb);
-        }
     }
 }
 
@@ -117,7 +97,12 @@ impl Widget for PushButtonWidget {
         self.text_widget.set_config(config, config_value.clone());
     }
 
-    fn handle_event(&mut self, injected: bool, event: CallbackEvent, _widget_store: Option<&Vec<WidgetContainer>>) -> Option<CallbackEvent> {
+    fn handle_event(
+        &mut self,
+        injected: bool,
+        event: CallbackEvent,
+        widget_store: Option<&Vec<WidgetContainer>>,
+    ) -> Option<CallbackEvent> {
         if !injected {
             match event {
                 CallbackEvent::MouseEntered { widget_id: _ } => {
@@ -151,11 +136,16 @@ impl Widget for PushButtonWidget {
                             self.draw_unhovered();
                             self.active = false;
 
-                            match _widget_store {
-                                Some(widgets) => {
-                                    self.click(widgets);
-                                },
-                                None => (),
+                            if self.get_callbacks().has_on_click() {
+                                match widget_store {
+                                    Some(widgets) => {
+                                        if let Some(mut cb) = self.get_callbacks().on_click.take() {
+                                            cb(self, widgets);
+                                            self.get_callbacks().on_click = Some(cb);
+                                        }
+                                    }
+                                    None => (),
+                                }
                             }
 
                             return Some(WidgetClicked { widget_id, button });
@@ -206,5 +196,9 @@ impl Widget for PushButtonWidget {
 
     fn get_drawable(&mut self) -> &mut dyn Drawable {
         self
+    }
+
+    fn get_callbacks(&mut self) -> &mut DefaultWidgetCallbacks {
+        &mut self.callbacks
     }
 }