+++ src/widget/text_widget.rs
@@ -21,6 +21,8 @@ use graphics::draw_state::DrawState;
 use graphics::*;
 use opengl_graphics::{GlGraphics, GlyphCache, TextureSettings};
 
+use crate::core::callbacks::*;
+use crate::core::widget_store::*;
 use crate::widget::config::*;
 use crate::widget::widget::*;
 
@@ -46,6 +48,7 @@ pub struct TextWidget {
     pub desired_width: i32,
     need_text_resize: bool,
     widget_id: i32,
+    callbacks: DefaultWidgetCallbacks,
 }
 
 impl TextWidget {
@@ -65,6 +68,7 @@ impl TextWidget {
             desired_width: 0 as i32,
             need_text_resize: true,
             widget_id: 0,
+            callbacks: DefaultWidgetCallbacks::new(),
         }
     }
 
@@ -166,6 +170,21 @@ impl Widget for TextWidget {
         self.widget_id
     }
 
+    fn handle_event(
+        &mut self,
+        injected: bool,
+        _event: CallbackEvent,
+        _widget_store: Option<&Vec<WidgetContainer>>,
+    ) -> Option<CallbackEvent> {
+        if !injected {}
+
+        None
+    }
+
+    fn handles_events(&mut self) -> bool {
+        true
+    }
+
     fn get_injectable_custom_events(&mut self) -> &mut dyn InjectableCustomEvents {
         self
     }
@@ -176,5 +195,9 @@ impl Widget for TextWidget {
 
     fn get_drawable(&mut self) -> &mut dyn Drawable {
         self
+    }
+
+    fn get_callbacks(&mut self) -> &mut DefaultWidgetCallbacks {
+        &mut self.callbacks
     }
 }