+++ src/widget/toggle_button_widget.rs
@@ -37,7 +37,7 @@ pub struct ToggleButtonWidget {
     selected: bool,
     active: bool,
     widget_id: i32,
-    on_click: Option<Box<dyn FnMut(&mut ToggleButtonWidget, bool)>>,
+    callbacks: DefaultWidgetCallbacks,
 }
 
 impl ToggleButtonWidget {
@@ -58,7 +58,7 @@ impl ToggleButtonWidget {
             selected: false,
             active: false,
             widget_id: 0,
-            on_click: None,
+            callbacks: DefaultWidgetCallbacks::new(),
         }
     }
 
@@ -88,26 +88,6 @@ impl ToggleButtonWidget {
         }
 
         self.invalidate();
-    }
-
-    /// Sets a callback closure that can be called when a click is registered for this
-    /// widget.
-    pub fn on_click<F>(&mut self, callback: F)
-    where
-        F: FnMut(&mut ToggleButtonWidget, bool) + 'static,
-    {
-        self.on_click = Some(Box::new(callback));
-    }
-
-    /// Calls the click `on_click` callback, if set.  Otherwise, ignored.  Sends a reference
-    /// of the current `Widget` object as a parameter, so this object can be modified when
-    /// a click is registered, if necessary.  Also indicates the state of the object, whether
-    /// or not the object has been toggled.
-    pub fn click(&mut self, state: bool) {
-        if let Some(mut cb) = self.on_click.take() {
-            cb(self, state);
-            self.on_click = Some(cb);
-        }
     }
 }
 
@@ -145,7 +125,12 @@ impl Widget for ToggleButtonWidget {
         }
     }
 
-    fn handle_event(&mut self, injected: bool, event: CallbackEvent, _widget_store: Option<&Vec<WidgetContainer>>) -> Option<CallbackEvent> {
+    fn handle_event(
+        &mut self,
+        injected: bool,
+        event: CallbackEvent,
+        widget_store: Option<&Vec<WidgetContainer>>,
+    ) -> Option<CallbackEvent> {
         if !injected {
             match event {
                 CallbackEvent::MouseEntered { widget_id: _ } => {
@@ -179,7 +164,21 @@ impl Widget for ToggleButtonWidget {
                             self.selected = !self.selected;
                             self.draw_unhovered();
                             self.active = false;
-                            self.click(self.selected);
+
+                            if self.get_callbacks().has_on_toggle() {
+                                let selected = self.selected;
+
+                                match widget_store {
+                                    Some(widgets) => {
+                                        if let Some(mut cb) = self.get_callbacks().on_toggle.take()
+                                        {
+                                            cb(self, selected, widgets);
+                                            self.get_callbacks().on_toggle = Some(cb);
+                                        }
+                                    }
+                                    None => (),
+                                }
+                            }
 
                             return Some(WidgetSelected {
                                 widget_id,
@@ -233,5 +232,9 @@ impl Widget for ToggleButtonWidget {
 
     fn get_drawable(&mut self) -> &mut dyn Drawable {
         self
+    }
+
+    fn get_callbacks(&mut self) -> &mut DefaultWidgetCallbacks {
+        &mut self.callbacks
     }
 }